Refactor Sum method to use LINQ for efficiency
The method `Sum` in the `SumOfMultiples` class has been updated to utilize LINQ, replacing the previous implementation that used a `HashSet` and nested loops. The new approach generates a range of numbers up to `max`, filters them for multiples of the provided values, and sums the results in a more concise manner.